Benchmark

+++ wiki/benchmark/locustfiles/cleanup.py
@@ -1,5 +1,5 @@
 from locust import HttpUser
-from config import AuthConfig
+from config import AuthConfig, CleanupConfig
 
 import actions
 
@@ -11,8 +11,14 @@ class Cleanup(HttpUser):
         
         # Delete generated items
         actions.deleteGeneratedAssetsAndPages(self)
-        actions.deleteGeneratedUsers(self)
-        actions.deleteGeneratedApiKeys(self)
+        
+        # Delete generated users
+        if CleanupConfig.delete_users:
+            actions.deleteGeneratedUsers(self)
+        
+        # Delete generated API Keys
+        if CleanupConfig.revoke_api_keys:
+            actions.deleteGeneratedApiKeys(self)
         
         # End session
         self.environment.runner.quit()
